loadTransaction un utils file. It also extracts method
