+++ src/pages/TransactionPage.vue
@@ -1,161 +1,35 @@
 <script lang="ts" setup>
 import { ref, watch } from 'vue';
 import { useI18n } from 'vue-i18n';
-import { ethers } from 'ethers';
 
 import { useRoute } from 'vue-router';
-import { contractManager, indexerApi } from 'src/boot/telosApi';
 
-import LogsViewer from 'components/Transaction/LogsViewer';
-import InternalTxns from 'components/Transaction/InternalTxns';
+import LogsViewer from 'components/Transaction/LogsViewer.vue';
+import InternalTxns from 'components/Transaction/InternalTxns.vue';
+import TransactionOverview from 'src/components/TransactionOverview.vue';
 
 const route = useRoute();
 
-import {
-    REVERT_FUNCTION_SELECTOR,
-    REVERT_PANIC_SELECTOR,
-    parseErrorMessage,
-} from 'src/lib/utils';
-
-
-import TransactionOverview from 'src/components/TransactionOverview.vue';
-import { EvmContractFunctionParameter, EvmTransaction, EvmTransactionLog } from 'src/antelope/types';
 import { EvmTransactionExtended } from 'src/types';
+import { loadTransaction } from 'src/lib/transaction-utils';
 
 const { t: $t } = useI18n();
 
-const REVERT_SELECTORS = [REVERT_FUNCTION_SELECTOR, REVERT_PANIC_SELECTOR];
 
 const defaultTab = 'overview';
 const tab = ref(defaultTab);
 const trxNotFound = ref(false);
 const hash = ref('');
-const errorMessage = ref<string | undefined>(undefined);
-const isContract = ref(false);
 const trx = ref<EvmTransactionExtended | null>(null);
 
-// functions
-const loadTransaction = async () => {
-    try {
-        const trxResponse = await indexerApi.get(`/transaction/${hash.value}?full=true&includeAbi=true`);
-        if (trxResponse.data.results.length === 0) {
-            trxNotFound.value = true;
-            return;
-        }
-        const aux = trxResponse.data.results[0] as EvmTransaction;
-        let logsArray: EvmTransactionLog[] = [];
-        if(aux.logs){
-            const fixedStr = aux.logs.replace('transaction_hash', 'transactionHash');
-            try {
-                logsArray = JSON.parse(fixedStr) as EvmTransactionLog[];
-            } catch (e) {
-                console.error('Error parsing logs', e);
-            }
-        }
-        const _trx:EvmTransactionExtended = {
-            ...aux,
-            gasUsedBn: ethers.BigNumber.from(aux.gasUsed),
-            gasLimitBn: ethers.BigNumber.from(aux.gasLimit),
-            valueBn: ethers.BigNumber.from(aux.value),
-            gasPriceBn: ethers.BigNumber.from(aux.gasPrice),
-            contract: undefined,
-            parsedTransaction: undefined,
-            functionParams: [],
-            logsArray,
-        };
-        await loadContract(_trx);
-        setErrorMessage();
-    } catch (e) {
-        trxNotFound.value = true;
-        return;
-    }
-};
 
-const loadContract = async (_trx: EvmTransactionExtended) => {
-    if (!_trx || _trx.input === '0x') {
-        trx.value = _trx;
-        return;
-    }
-    _trx.contract = await contractManager.getContract(_trx.to?.toLowerCase());
-    if (!_trx.contract) {
-        trx.value = _trx;
-        return;
-    }
 
-    _trx.parsedTransaction = await contractManager.parseContractTransaction(
-        _trx,
-        _trx.input,
-        _trx.contract,
-    );
-
-    _trx.functionParams = getFunctionParams();
-
-    trx.value = _trx;
-    isContract.value = true;
-};
-
-const setErrorMessage = async () => {
-    if (!trx.value || trx.value.status === '0x1') {
-        return;
-    }
-    await loadErrorMessage();
-};
-
-const loadErrorMessage = async () => {
-    if (!trx.value) {
-        return;
-    }
-
-    if(trx.value.output && REVERT_SELECTORS.includes(trx.value.output.slice(0, 10))){
-        errorMessage.value = parseErrorMessage(trx.value.output);
-        return;
-    }
-    const response = await indexerApi.get(
-        `/transaction/${hash.value}/internal`,
-    );
-    const intrxs = response.data.results;
-    for(let i = 0; i < intrxs.length; i++){
-        const intrx = intrxs[i];
-
-        if (!intrx.result?.output) {
-            continue;
-        }
-
-        const output = (REVERT_SELECTORS.includes(intrx.result?.output.slice(0, 10)))
-            ? intrx.result?.output
-            : trx.value.output
-        ;
-        if(intrx.error !== null){
-            errorMessage.value = (REVERT_SELECTORS.includes(output?.slice(0, 10)))
-                ? parseErrorMessage(output)
-                : intrx.error
-            ;
-            return;
-        }
-    }
-};
-const getFunctionParams = () => {
-    if (!trx.value?.parsedTransaction) {
-        return [];
-    }
-    const args:EvmContractFunctionParameter[] = [];
-    trx.value?.parsedTransaction.functionFragment.inputs.forEach((input, i) => {
-        args.push({
-            name: input.name,
-            type: input.type,
-            arrayChildren: (input.arrayChildren !== null) ? input.arrayChildren.type : false,
-            value:  trx.value?.parsedTransaction?.args[i],
-        });
-    });
-    return args;
-};
-
-watch(() => route.params.hash, (newValue) => {
+watch(() => route.params.hash, async (newValue) => {
     if (!newValue || hash.value === newValue) {
         return;
     }
     hash.value = typeof newValue === 'string' ? newValue : newValue[0];
-    loadTransaction();
+    trx.value = await loadTransaction(hash.value);
 }, { immediate: true });
